sdfgsdgsd

+++ ncmanstreet/src/main/java/com/rcs/nchumanity/entity/model/EmergencyInfo.java
@@ -14,9 +14,9 @@ public class EmergencyInfo implements Serializable {
 
     private String content;
 
-    private Float longitude;
+    private Double longitude;
 
-    private Float latitude;
+    private Double latitude;
 
     private Integer readCount;
 
@@ -82,19 +82,19 @@ public class EmergencyInfo implements Serializable {
         this.content = content == null ? null : content.trim();
     }
 
-    public Float getLongitude() {
+    public Double  getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
